+++ test/test_lowlevel.py
@@ -740,11 +740,11 @@ class TestMULTIDIMARRAYS:
 
         cppyy.cppdef(r"""\
         namespace StringArray {
-           char str_array[3][8] = {"s1\0", "s23\0", "s456\0"};
+           char str_array[3][7] = {"s1\0", "s23\0", "s456\0"};
         }""")
 
         ns = cppyy.gbl.StringArray
 
         for i, v in enumerate(("s1", "s23", "s456")):
-            assert len(ns.str_array[i]) == 8
-            assert str(ns.str_array[i])[:len(v)] == v
+            assert len(ns.str_array[i]) == 7
+            assert ns.str_array[i].as_string() == v